prod_agent/requirements.txt
requests
boto3
strands-agents 
strands-agents-tools
bedrock-agentcore
uv